refactor: move thread runtime message repair
