+++ backend/thread_runtime/history.py
@@ -6,7 +6,7 @@ from collections.abc import Awaitable, Callable
 from dataclasses import dataclass
 from typing import Any
 
-from backend.web.services.thread_message_interruption_service import repair_interrupted_tool_call_messages
+from backend.thread_runtime.interruption import repair_interrupted_tool_call_messages
 from backend.web.utils.serializers import extract_text_content
